Got a lot done
- Worked with Android's new permission system in order to request
permission for Internet access.
- Removed some verbose logging once I saw JSON was fine.
- Changed POSTER_PATH to take into account double slash URL.
- Tried to get the grid to look good.

The images are loading into the app, but they aren't appearing the right
size and they aren't filling the grid space (so it looks just like a
list). I think I need to force the width of the movie_poster_item.xml to
be half the screen. Will work on that tomorrow.